Apply the same pattern of requiring a file to be loaded to all other tools
