+++ src/lib/index.ts
@@ -6,6 +6,7 @@ export { default as IconButton } from "./IconButton.svelte";
 export { default as Loading } from "./Loading.svelte";
 export { default as Modal } from "./Modal.svelte";
 export { default as PrevNext } from "./PrevNext.svelte";
+export { default as ToolLayout } from "./ToolLayout.svelte";
 export { default as YesNo } from "./YesNo.svelte";
 
 export function pairs(list: string[]): [string, string][] {
