+++ src/routes/+page.svelte
@@ -2,14 +2,18 @@
   import { base } from "$app/paths";
 </script>
 
-<p>Design assistance</p>
-<ul>
-  <li><a href="{base}/cross_section">Route Cross Section Tool</a></li>
-</ul>
+<div class="govuk-width-container">
+  <p>Design assistance</p>
+  <ul>
+    <li><a href="{base}/cross_section">Route Cross Section Tool</a></li>
+  </ul>
 
-<p>Scheme review</p>
-<ul>
-  <li><a href="{base}/area_check">Area Check</a></li>
-  <li><a href="{base}/route_check">Route Check</a></li>
-  <li><a href="{base}/planning">Planning Application Assessment Toolkit</a></li>
-</ul>
+  <p>Scheme review</p>
+  <ul>
+    <li><a href="{base}/area_check">Area Check</a></li>
+    <li><a href="{base}/route_check">Route Check</a></li>
+    <li>
+      <a href="{base}/planning">Planning Application Assessment Toolkit</a>
+    </li>
+  </ul>
+</div>
