+++ src/routes/area_check/+layout.svelte
@@ -1,27 +1,8 @@
 <script lang="ts">
-  import { currentFile } from "./data";
-  import { base } from "$app/paths";
-  import { NavHeader, NavFooter, getTitle } from "$lib/nav";
-  import { page } from "$app/stores";
-  import folderUrl from "$lib/assets/images/folder.svg?url";
+  import { state, currentFile, files } from "./data";
+  import { ToolLayout } from "$lib";
 </script>
 
-<div>
-  <img src={folderUrl} alt="Manage files" style="vertical-align: middle;" />
-  <a href="{base}/area_check/files">Manage files</a>
-  <span class="govuk-body" style="margin-left: 8px;">
-    Editing file <u>{$currentFile}</u>
-  </span>
-</div>
-
-<NavHeader routeCheckType="" />
-
-<hr />
-
-<h1 style="text-align: center">{getTitle($page.url.pathname)}</h1>
-
-<slot />
-
-<hr />
-
-<NavFooter routeCheckType="" />
+<ToolLayout {state} {currentFile} {files} tool="area_check" routeCheckType="">
+  <slot />
+</ToolLayout>
