--- src/routes/area_check/files/+layout@.svelte
@@ -1,3 +1,0 @@
-<!-- Avoid area_check/layout.svelte, because we don't need a link to this page -->
-<hr />
-<slot />
