--- src/routes/cross_section/+page.svelte
@@ -1,19 +1,0 @@
-<script lang="ts">
-  import { base } from "$app/paths";
-  import { ExternalLink } from "$lib";
-</script>
-
-<p>
-  This is an experimental version of a <ExternalLink
-    href="https://www.gov.uk/government/publications/active-travel-england-design-assistance-tools"
-  >
-    design assistance tool
-  </ExternalLink>.
-</p>
-
-<ol>
-  <li><a href="{base}/cross_section/summary">Summary of Scheme</a></li>
-  <li><a href="{base}/cross_section/proposed">Proposed Cross-Sections</a></li>
-  <li><a href="{base}/cross_section/check">Cross-Sections Check</a></li>
-  <li><a href="{base}/cross_section/results_summary">ATE Summary</a></li>
-</ol>
