--- src/routes/cross_section/files/+layout@.svelte
@@ -1,3 +1,0 @@
-<!-- Avoid cross_section/layout.svelte, because we don't need a link to this page -->
-<hr />
-<slot />
