--- src/routes/planning/+page.svelte
@@ -1,11 +1,0 @@
-<script lang="ts">
-  import { base } from "$app/paths";
-</script>
-
-<p>This is an experimental version of an internal tool.</p>
-
-<ol>
-  <li><a href="{base}/planning/app_details">Application details</a></li>
-  <li><a href="{base}/planning/scorecard">Scorecard</a></li>
-  <li><a href="{base}/planning/report">Report</a></li>
-</ol>
