--- src/routes/planning/files/+layout@.svelte
@@ -1,3 +1,0 @@
-<!-- Avoid planning/layout.svelte, because we don't need a link to this page -->
-<hr />
-<slot />
