--- src/routes/planning/files/+page.svelte
@@ -1,6 +1,0 @@
-<script lang="ts">
-  import { FileManager } from "$lib/files";
-  import { files, currentFile, state } from "../data";
-</script>
-
-<FileManager {files} {currentFile} {state} />
