+++ src/routes/planning/nav/+page.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+  import { base } from "$app/paths";
+</script>
+
+<div class="govuk-width-container">
+  <p>This is an experimental version of an internal tool.</p>
+
+  <ol>
+    <li><a href="{base}/planning/app_details">Application details</a></li>
+    <li><a href="{base}/planning/scorecard">Scorecard</a></li>
+    <li><a href="{base}/planning/report">Report</a></li>
+  </ol>
+</div>
